multi-PCM W9: daemon side of per-PCM Reset (sentinel convention)
DriverManager::reset() widens its parameter from uint8_t to int32_t to carry the W9 reset-all sentinel; the two init sequences (real + fake driver) switch reset(0) -> reset(-1) so startup still gets a full wipe under the new kernel convention (pcm_id < 0 = all, >= 0 = that PCM only). Wire payload unchanged (already int32_t).

Pairs with the lkm W9 MT_ALSA_Msg_Reset dispatch.

+++ daemon/driver_manager.hpp
@@ -74,7 +74,7 @@ class DriverManager : public DriverHandler {
   std::error_code hello();
   std::error_code start();
   std::error_code stop();
-  std::error_code reset(uint8_t pcm_id);
+  std::error_code reset(int32_t pcm_id);  // pcm_id < 0 = all PCMs (clean slate); >= 0 = that PCM only (W9)
   std::error_code bye();
 
   void on_command_done(enum MT_ALSA_msg_id id,
